# [RSS] Add manual performance tests that run git clone + bazel for more realistic results