@maggie-lou 2023-10-20
> LGTM, but any idea why the build on a clean runner takes twice as long w/ RSS enabled? Seems like the main slowness would be from Pause() but it looks like we're reporting the time before we call Pause()

The numbers can vary quite a bit between runs. When I rerun it a couple times, the clean runner with RSS enabled vs disabled takes about the same time actually. I'll update the comment so that it's not as misleading